# Ability to specify different distance function