--- a/examples/lloyd.rs
+++ b/examples/lloyd.rs
@@ -8,7 +8,7 @@ fn main() {
     samples.iter_mut().for_each(|v| *v = rand::random());
 
     // Calculate kmeans, using kmean++ as initialization-method
-    let kmean: KMeans<f64, 8> = KMeans::new(samples, sample_cnt, sample_dims);
+    let kmean: KMeans<f64, 8, _> = KMeans::new(samples, sample_cnt, sample_dims, EuclideanDistance);
     let result = kmean.kmeans_lloyd(k, max_iter, KMeans::init_kmeanplusplus, &KMeansConfig::default());
 
     println!("Centroids: {:?}", result.centroids);