--- a/src/distances/mod.rs
+++ b/src/distances/mod.rs
@@ -1,0 +1,5 @@
+mod euclideandistance;
+mod histogramdistance;
+
+pub use euclideandistance::EuclideanDistance;
+pub use histogramdistance::HistogramDistance;