--- a/src/inits/randompartition.rs
+++ b/src/inits/randompartition.rs
@@ -1,14 +1,16 @@
+use crate::api::DistanceFunction;
 use crate::memory::*;
 use crate::{KMeans, KMeansConfig, KMeansState};
 use rand::prelude::*;
 use std::simd::{LaneCount, Simd, SupportedLaneCount};
 
 #[inline(always)]
-pub fn calculate<T, const LANES: usize>(kmean: &KMeans<T, LANES>, state: &mut KMeansState<T>, config: &KMeansConfig<'_, T>)
+pub fn calculate<T, const LANES: usize, D>(kmean: &KMeans<T, LANES, D>, state: &mut KMeansState<T>, config: &KMeansConfig<'_, T>)
 where
     T: Primitive,
     LaneCount<LANES>: SupportedLaneCount,
     Simd<T, LANES>: SupportedSimdArray<T, LANES>,
+    D: DistanceFunction<T, LANES>,
 {
     let (assignments, centroids, centroid_frequency, k) =
         (&mut state.assignments, &mut state.centroids, &mut state.centroid_frequency, state.k);