--- a/src/inits/randomsample.rs
+++ b/src/inits/randomsample.rs
@@ -1,3 +1,4 @@
+use crate::api::DistanceFunction;
 use crate::memory::*;
 use crate::{KMeans, KMeansConfig, KMeansState};
 use rand::prelude::*;
@@ -5,11 +6,12 @@ use std::ops::DerefMut;
 use std::simd::{LaneCount, Simd, SupportedLaneCount};
 
 #[inline(always)]
-pub fn calculate<T, const LANES: usize>(kmean: &KMeans<T, LANES>, state: &mut KMeansState<T>, config: &KMeansConfig<'_, T>)
+pub fn calculate<T, const LANES: usize, D>(kmean: &KMeans<T, LANES, D>, state: &mut KMeansState<T>, config: &KMeansConfig<'_, T>)
 where
     T: Primitive,
     LaneCount<LANES>: SupportedLaneCount,
     Simd<T, LANES>: SupportedSimdArray<T, LANES>,
+    D: DistanceFunction<T, LANES>,
 {
     kmean
         .p_samples