--- a/src/lib.rs
+++ b/src/lib.rs
@@ -42,7 +42,7 @@
 //!
 //! // Calculate kmeans, using kmean++ as initialization-method
 //! // KMeans<_, 8> specifies to use f64 SIMD vectors with 8 lanes (e.g. AVX512)
-//! let kmean: KMeans<_, 8> = KMeans::new(samples, sample_cnt, sample_dims);
+//! let kmean: KMeans<_, 8, _> = KMeans::new(samples, sample_cnt, sample_dims, EuclideanDistance);
 //! let result = kmean.kmeans_lloyd(k, max_iter, KMeans::init_kmeanplusplus, &KMeansConfig::default());
 //!
 //! println!("Centroids: {:?}", result.centroids);
@@ -69,7 +69,7 @@
 //!
 //! // Calculate kmeans, using kmean++ as initialization-method
 //! // KMeans<_, 8> specifies to use f64 SIMD vectors with 8 lanes (e.g. AVX512)
-//! let kmean: KMeans<_, 8> = KMeans::new(samples, sample_cnt, sample_dims);
+//! let kmean: KMeans<_, 8, _> = KMeans::new(samples, sample_cnt, sample_dims, EuclideanDistance);
 //! let result = kmean.kmeans_minibatch(4, k, max_iter, KMeans::init_random_sample, &conf);
 //!
 //! println!("Centroids: {:?}", result.centroids);
@@ -97,17 +97,20 @@ extern crate test;
 mod helpers;
 mod abort_strategy;
 mod api;
+mod distances;
 mod inits;
 mod memory;
 mod variants;
 
 pub use abort_strategy::AbortStrategy;
-pub use api::{KMeans, KMeansConfig, KMeansConfigBuilder, KMeansState};
+pub use api::{DistanceFunction, KMeans, KMeansConfig, KMeansConfigBuilder, KMeansState};
+pub use distances::{EuclideanDistance, HistogramDistance};
 pub use memory::Primitive;
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use distances::EuclideanDistance;
     use memory::SupportedSimdArray;
     use rand::prelude::*;
     use std::simd::{LaneCount, Simd, SupportedLaneCount};
@@ -141,7 +144,7 @@ mod tests {
         let mut rnd = rand::rngs::StdRng::seed_from_u64(1337);
         let mut samples = vec![T::zero(); sample_cnt * sample_dims];
         samples.iter_mut().for_each(|v| *v = rnd.gen_range(T::zero()..T::one()));
-        let kmean: KMeans<T, LANES> = KMeans::new(samples, sample_cnt, sample_dims);
+        let kmean: KMeans<T, LANES, _> = KMeans::new(samples, sample_cnt, sample_dims, EuclideanDistance);
         let conf = KMeansConfig::build().random_generator(rnd).build();
         b.iter(|| kmean.kmeans_lloyd(k, max_iter, KMeans::init_kmeanplusplus, &conf));
     }
@@ -173,7 +176,7 @@ mod tests {
         let mut rnd = rand::rngs::StdRng::seed_from_u64(1337);
         let mut samples = vec![T::zero(); sample_cnt * sample_dims];
         samples.iter_mut().for_each(|v| *v = rnd.gen_range(T::zero()..T::one()));
-        let kmean = KMeans::new(samples, sample_cnt, sample_dims);
+        let kmean = KMeans::new(samples, sample_cnt, sample_dims, EuclideanDistance);
         let conf = KMeansConfig::build().random_generator(rnd).build();
         b.iter(|| kmean.kmeans_minibatch(batch_size, k, max_iter, KMeans::init_random_sample, &conf));
     }