Would have liked to not break the API, but couldn't figure out how to set EuclideanDistance as the default one