@seijikun 2024-10-22
I think that feature is well worth the API breakage. Thanks!

Can you please:
- Fix the unit tests I get a lot of: Missing 3rd type parameter errors.
- Move the distance functions into a folder structure like `inits` and `variants`
- Add `#[inline(always)]` on `EuclideanDistance::distance()` and `HistogramDistance::distance()`
- Run `rustfmt --edition 2021 ./src/lib.rs`

How is the performance difference between before and after this change?