@reinismu 2024-11-06
It seems it didn't change. 

Before
```
test api::tests::distance_matrix_calculation_benchmark_f32x16 ... bench:   8,632,870.85 ns/iter (+/- 856,820.38)
test api::tests::distance_matrix_calculation_benchmark_f32x2  ... bench:   5,084,460.45 ns/iter (+/- 446,414.39)
test api::tests::distance_matrix_calculation_benchmark_f32x4  ... bench:   5,002,356.30 ns/iter (+/- 325,506.24)
test api::tests::distance_matrix_calculation_benchmark_f32x8  ... bench:   5,628,987.55 ns/iter (+/- 288,415.62)
test api::tests::distance_matrix_calculation_benchmark_f64x2  ... bench:   5,237,735.40 ns/iter (+/- 300,572.49)
test api::tests::distance_matrix_calculation_benchmark_f64x4  ... bench:   5,807,877.10 ns/iter (+/- 356,531.25)
test api::tests::distance_matrix_calculation_benchmark_f64x8  ... bench:   8,337,633.30 ns/iter (+/- 440,314.17)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x16   ... bench:  10,455,512.50 ns/iter (+/- 410,819.59)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x2    ... bench:  10,324,604.20 ns/iter (+/- 232,773.84)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x4    ... bench:   8,153,587.50 ns/iter (+/- 239,791.28)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x8    ... bench:   8,390,075.00 ns/iter (+/- 722,864.14)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f64x2    ... bench:  11,214,166.70 ns/iter (+/- 731,121.57)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f64x4    ... bench:   9,701,468.70 ns/iter (+/- 341,047.92)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f64x8    ... bench:  10,609,533.40 ns/iter (+/- 693,425.32)
test tests::complete_benchmark_lloyd_big_f32x8                ... bench:   7,506,354.20 ns/iter (+/- 213,025.25)
test tests::complete_benchmark_lloyd_big_f64x8                ... bench:   8,257,418.75 ns/iter (+/- 525,992.48)
test tests::complete_benchmark_lloyd_huge_f32x8               ... bench:  46,334,391.70 ns/iter (+/- 2,430,593.89)
test tests::complete_benchmark_lloyd_huge_f64x8               ... bench:  73,769,920.80 ns/iter (+/- 5,943,434.07)
test tests::complete_benchmark_lloyd_mid_f32x8                ... bench:   9,390,227.10 ns/iter (+/- 531,478.14)
test tests::complete_benchmark_lloyd_mid_f64x8                ... bench:  12,776,008.30 ns/iter (+/- 786,151.22)
test tests::complete_benchmark_lloyd_small_f32x8              ... bench:   6,841,452.05 ns/iter (+/- 155,411.66)
test tests::complete_benchmark_lloyd_small_f64x8              ... bench:   9,180,383.35 ns/iter (+/- 392,392.28)
test tests::complete_benchmark_minibatch_big_f32x8            ... bench:   6,046,856.20 ns/iter (+/- 1,130,595.90)
test tests::complete_benchmark_minibatch_big_f64x8            ... bench:   6,754,314.60 ns/iter (+/- 1,471,971.86)
test tests::complete_benchmark_minibatch_huge_f32x8           ... bench:  25,392,000.00 ns/iter (+/- 3,100,102.03)
test tests::complete_benchmark_minibatch_huge_f64x8           ... bench:  38,191,762.50 ns/iter (+/- 7,328,353.05)
test tests::complete_benchmark_minibatch_mid_f32x8            ... bench:   3,208,257.27 ns/iter (+/- 493,290.73)
test tests::complete_benchmark_minibatch_mid_f64x8            ... bench:   4,623,425.00 ns/iter (+/- 1,097,594.44)
test tests::complete_benchmark_minibatch_small_f32x8          ... bench:   2,445,235.40 ns/iter (+/- 404,960.29)
test tests::complete_benchmark_minibatch_small_f64x8          ... bench:   3,475,285.40 ns/iter (+/- 779,652.75)

test result: ok. 0 passed; 0 failed; 12 ignored; 30 measured; 0 filtered out; finished in 205.84s
```

With distance changed
```
test api::tests::distance_matrix_calculation_benchmark_f32x16 ... bench:   8,298,629.20 ns/iter (+/- 245,060.35)
test api::tests::distance_matrix_calculation_benchmark_f32x2  ... bench:   5,144,660.40 ns/iter (+/- 2,273,019.33)
test api::tests::distance_matrix_calculation_benchmark_f32x4  ... bench:   4,986,125.05 ns/iter (+/- 370,299.79)
test api::tests::distance_matrix_calculation_benchmark_f32x8  ... bench:   5,571,014.55 ns/iter (+/- 333,595.20)
test api::tests::distance_matrix_calculation_benchmark_f64x2  ... bench:   5,220,945.85 ns/iter (+/- 364,674.98)
test api::tests::distance_matrix_calculation_benchmark_f64x4  ... bench:   5,724,989.60 ns/iter (+/- 256,319.36)
test api::tests::distance_matrix_calculation_benchmark_f64x8  ... bench:   8,292,858.30 ns/iter (+/- 457,210.47)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x16   ... bench:  10,496,183.30 ns/iter (+/- 669,229.65)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x2    ... bench:  10,282,737.50 ns/iter (+/- 305,680.44)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x4    ... bench:   8,196,781.25 ns/iter (+/- 370,984.22)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f32x8    ... bench:   8,399,770.85 ns/iter (+/- 622,216.03)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f64x2    ... bench:  11,271,162.50 ns/iter (+/- 413,270.52)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f64x4    ... bench:  10,973,816.70 ns/iter (+/- 688,134.62)
test inits::kmeanplusplus::tests::init_kmeanplusplus_f64x8    ... bench:  10,624,775.00 ns/iter (+/- 222,987.18)
test tests::complete_benchmark_lloyd_big_f32x8                ... bench:   7,513,847.90 ns/iter (+/- 387,775.01)
test tests::complete_benchmark_lloyd_big_f64x8                ... bench:   8,294,331.25 ns/iter (+/- 399,443.50)
test tests::complete_benchmark_lloyd_huge_f32x8               ... bench:  46,332,987.60 ns/iter (+/- 442,971.71)
test tests::complete_benchmark_lloyd_huge_f64x8               ... bench:  72,555,283.30 ns/iter (+/- 2,842,621.74)
test tests::complete_benchmark_lloyd_mid_f32x8                ... bench:   9,611,914.55 ns/iter (+/- 663,866.84)
test tests::complete_benchmark_lloyd_mid_f64x8                ... bench:  12,840,895.80 ns/iter (+/- 591,495.03)
test tests::complete_benchmark_lloyd_small_f32x8              ... bench:   6,920,958.35 ns/iter (+/- 1,238,978.76)
test tests::complete_benchmark_lloyd_small_f64x8              ... bench:  10,411,029.10 ns/iter (+/- 2,812,717.74)
test tests::complete_benchmark_minibatch_big_f32x8            ... bench:   6,110,493.75 ns/iter (+/- 1,002,475.68)
test tests::complete_benchmark_minibatch_big_f64x8            ... bench:   6,708,708.30 ns/iter (+/- 1,641,328.94)
test tests::complete_benchmark_minibatch_huge_f32x8           ... bench:  25,409,916.60 ns/iter (+/- 3,087,121.98)
test tests::complete_benchmark_minibatch_huge_f64x8           ... bench:  36,239,583.30 ns/iter (+/- 5,131,284.97)
test tests::complete_benchmark_minibatch_mid_f32x8            ... bench:   3,217,296.85 ns/iter (+/- 487,658.92)
test tests::complete_benchmark_minibatch_mid_f64x8            ... bench:   4,609,522.90 ns/iter (+/- 1,110,200.40)
test tests::complete_benchmark_minibatch_small_f32x8          ... bench:   2,444,686.45 ns/iter (+/- 328,333.44)
test tests::complete_benchmark_minibatch_small_f64x8          ... bench:   3,491,535.40 ns/iter (+/- 817,114.81)

test result: ok. 0 passed; 0 failed; 12 ignored; 30 measured; 0 filtered out; finished in 189.27s
```