@seijikun 2024-11-16
Awesome, thanks!
Gonna bump the major version number due to the API change.