@seijikun 2024-11-16
I just releasted v1.0.0 containing your changes